# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GameSessionStatus =
  (typeof GameSessionStatus)[keyof typeof GameSessionStatus];

export const GameSessionStatus = {
  active: "active",
  won: "won",
  lost: "lost",
} as const;

export interface GameSession {
  sessionId: string;
  status: GameSessionStatus;
  guessCount: number;
  maxGuesses: number;
  minRange: number;
  maxRange: number;
  guesses: number[];
  /** @nullable */
  lastHint?: string | null;
}

export interface GuessInput {
  sessionId: string;
  guess: number;
}

export type GuessResultHint =
  (typeof GuessResultHint)[keyof typeof GuessResultHint];

export const GuessResultHint = {
  too_low: "too_low",
  too_high: "too_high",
  correct: "correct",
  lost: "lost",
} as const;

export type GuessResultStatus =
  (typeof GuessResultStatus)[keyof typeof GuessResultStatus];

export const GuessResultStatus = {
  active: "active",
  won: "won",
  lost: "lost",
} as const;

export interface GuessResult {
  hint: GuessResultHint;
  status: GuessResultStatus;
  guessCount: number;
  maxGuesses: number;
  guesses: number[];
  /** @nullable */
  secretNumber?: number | null;
}

export interface LeaderboardEntry {
  sessionId: string;
  guessCount: number;
  minRange: number;
  maxRange: number;
}