# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: game
    description: Number guessing game
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /game/start:
    post:
      operationId: startGame
      tags: [game]
      summary: Start a new game
      description: Starts a new number guessing game and returns a session ID
      responses:
        "201":
          description: Game started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"

  /game/guess:
    post:
      operationId: submitGuess
      tags: [game]
      summary: Submit a guess
      description: Submit a number guess for an active game session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuessInput"
      responses:
        "200":
          description: Guess result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuessResult"
        "404":
          description: Session not found

  /game/state/{sessionId}:
    get:
      operationId: getGameState
      tags: [game]
      summary: Get current game state
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Game state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
        "404":
          description: Session not found

  /game/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [game]
      summary: Get top scores
      description: Returns top completed games sorted by fewest guesses
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    GameSession:
      type: object
      required:
        - sessionId
        - status
        - guessCount
        - maxGuesses
        - minRange
        - maxRange
        - guesses
      properties:
        sessionId:
          type: string
        status:
          type: string
          enum: [active, won, lost]
        guessCount:
          type: integer
        maxGuesses:
          type: integer
        minRange:
          type: integer
        maxRange:
          type: integer
        guesses:
          type: array
          items:
            type: integer
        lastHint:
          type: ["string", "null"]

    GuessInput:
      type: object
      required:
        - sessionId
        - guess
      properties:
        sessionId:
          type: string
        guess:
          type: integer

    GuessResult:
      type: object
      required:
        - hint
        - status
        - guessCount
        - maxGuesses
        - guesses
      properties:
        hint:
          type: string
          enum: [too_low, too_high, correct, lost]
        status:
          type: string
          enum: [active, won, lost]
        guessCount:
          type: integer
        maxGuesses:
          type: integer
        guesses:
          type: array
          items:
            type: integer
        secretNumber:
          type: ["integer", "null"]

    LeaderboardEntry:
      type: object
      required:
        - sessionId
        - guessCount
        - maxRange
        - minRange
      properties:
        sessionId:
          type: string
        guessCount:
          type: integer
        minRange:
          type: integer
        maxRange:
          type: integer