# lib/api-zod/src/generated/types/gameSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameSessionStatus } from "./gameSessionStatus";

export interface GameSession {
  sessionId: string;
  status: GameSessionStatus;
  guessCount: number;
  maxGuesses: number;
  minRange: number;
  maxRange: number;
  guesses: number[];
  /** @nullable */
  lastHint?: string | null;
}