# lib/api-zod/src/generated/types/guessResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuessResultHint } from "./guessResultHint";
import type { GuessResultStatus } from "./guessResultStatus";

export interface GuessResult {
  hint: GuessResultHint;
  status: GuessResultStatus;
  guessCount: number;
  maxGuesses: number;
  guesses: number[];
  /** @nullable */
  secretNumber?: number | null;
}